Asking for user input for variables
This script asks the user for input and uses it as a parameter for the
executable. It could be improved by asking for the user to select from a
list of options for these particular variables. Asking the user for a
string probably isn't ideal, but it's my first crack at it.