If statements
This is a simple rock paper scissors game using If statements. Once I
got the basic functionality down, I cleaned up the formatting a bit with
some blank lines. It should tell if you if you pick an invalid INT. If
you type something that is not an INT, it will just error out.